Post, Delete,Put Requests are Done

+++ restapi/src/main/java/com/viscoloveapi01/Pillows02BeanRepository.java
@@ -2,6 +2,6 @@ package com.viscoloveapi01;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Pillows02BeanRepository extends JpaRepository<Pillows01Bean, Long> {
+public interface Pillows02BeanRepository extends JpaRepository<Pillows01Bean, Integer> {
 
 }
