+++ restapi/src/main/java/com/viscoloveapi02/Visco02BeanRepository.java
@@ -2,7 +2,7 @@ package com.viscoloveapi02;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Visco02BeanRepository extends JpaRepository<Visco01Bean, Long> {
+public interface Visco02BeanRepository extends JpaRepository<Visco01Bean, Integer> {
 
 
 }
